refactor: 重构文件结构，优化模块导入路径并添加新的 AlphaTab 相关功能
