+++ src/alphatab/AlphaTabSettingsHelper.ts
